Add the ability to start workers from examples.
This makes it much easier to experiment with worker options/count, rather than
having to fire up external workers in docker/kubernetes etc.

It will also enable the use of checks and thresholds in k6 as the metrics now
go through the k6 infrastructure.

This does not prevent external workers from being used (that will always be a
requirement in order to load/stress test real infrastructure at intended
scale).